Suggested full example tweak
Not saying this is definitely better, but if I were to write this function I'd refactor it like this, so it can take as many NUM URI's as I want to pass it, and loops through them.

I've not tested the code so you'd need to check it if you prefer this route. Totally fine if you think it's simpler to not include a loop here, up to you.